[compiler] Move IVs into the CreateLoopOpts struct
This deprioritises the IVs parameter of `createLoop`, in the case that
users don't wish to set it and are happy with the default of 'none'.

It also allows us to more cohesively provide a second options field
which allows users to set the IV names, rather than a) adding an
additional parameter to `createLoop` or b) having related values split
across `createLoop` and `CreateLoopOpts`.