[host] Fix move-after-move errors
The use of `llvm::consumeError` is just to mark the error as 'handled' - it
doesn't actually report the error in any way. If there's a callback
we have reported *and* handled the error, so using `llvm::consumeError`
is more correctly placed in an 'else' block.

This fixes potential move-after-move errors when a compilation error
happens and the callback doesn't halt execution at that point.