Support all LLVM versions with same maximum major version
This should reduce the noise required to support newly released LLVM
versions, as we don't have to explicitly opt in to all minor and patch
releases. This assumes that all newer minor/patch versions are strict
improvements on previous versions.

The minimum version has been kept fully qualified as it is feasible that
we may require a minor/patch versions as a minimum.