[spirv-ll] Ensure that GroupUniformArithmeticKHR nodes have results
This fixes a crash exposed by hacking the SYCL-CTS somewhat.

Note that due to our coarse-grained tracking of `spirv-as` versions for
test purposes, this test is pessimistically run on any v2024+ version
(i.e., never...), when in fact we want the as-yet unreleased v2023.5.
This will be fixed up in the near future as it's a pre-existing issue.