[spirv-ll] Improve spirv-as version tracking for test features
This improves our previous tracking of `spirv-as` versions which was
based soley on their release year, ignoring the point release. This
coarse grained system meant that we had to conditionally enable spirv-ll
tests on versions far newer than the oldest version which supports them,
e.g., a feature introduce in `v2022.1` couldn't be tested until
`v2022.0` was released.

It also now tracks `-dev` versions, which are packaged on some
distributions as the 'official' `spirv-as` binary. We treat these as the
previous point release, as in testing we have discovered that
`v2022.1-dev` doesn't have some features present in `v2022.1`. Thus a
`dev` version decrements the point release by 1, and if that would
result in a negative version, we decrement the year instead, and set the
point to 9. We choose 9 somewhat arbitrarily: the tools have never
released such a high point number. For example, we internally track
versions thus:

    v2022.1-dev -> v2022.0
    v2022.0-dev -> v2021.9

We also print this to the console during CMake configuration, for better
clarity.

Note that we must do this both in CMake (for offline configurations) and
in LIT (for online configurations), but the logic is the same.

With this internal information in place, we now add LIT features for
point releases. We add all point releases 0..9 for previous years, so
v2022.1 would add the following features:

    v2021.0+, v2021.1+, ..., v2021.9+, v2022.0+, v2022.1+

It similarily adds the same ten 'point' features 0..9 for all years from
2016 until the 'current' year. This is another reason for choosing 9 as
the highest point release: we don't want to add an unreasonably high
number of features (e.g., 99 per year) as it bloats the feature list and
makes more improtant features harder to spot during development.

With these new features, we can retroactively enables some LIT tests for
older `spirv-as` versions.