Fix for SPIR-V translation to IR for OpPtrAccessChain
In some cases the indexes field for OpPtrAccessChain can be empty.
The translation code currently assumes this not to be the case and
checks for element decorations.

This fixes the empty index case by not checking the decorations in this
case. There are some issues with OpMemberDecorate, so testing of the
path with additional arguments requiring decorations will not be tested here.

Also some previous clang-format issues have been resolved.