[passes] Handle alignments of structs correctly with local memory
This pass changes accesses to local memory (in the form of globals
in address space 3) to offsets into a single struct passed into
the function.

However, when calculating the layout of this struct, it incorrectly
assumed that the size of a type was equal to its alignment (with
some special handling for arrays). This patch changes it so that
the datalayout is queried directly for the size of the type.