[compiler] Represent samplers/events internally as size_t
This fixes a bug in LLVM 17 where, on 64-bit targets, the `target` types
used to represent samplers and events have an implicit size/alignment of
8 bytes (as their underlying "layout type" is `ptr`). This means that it
is generally invalid to replace them with `i32`, as this may break
existing alignment/size assumptions that the rest of the module has made
on the access of the 'old' types.

Instead, we choose to replace `target` types with `size_t` (in practice
`i32` or `i64` depending on the SPIR triple). This works more or less
the same way on 32-bit targets.

On 64-bit targets, however, the runtime passes samplers to kernels as
8-byte values. These values follow the existing munging via pointers on
LLVM 16, but are eventually zero-extended or truncated to `i32` when
they interface with existing APIs that expect `i32` - such as `libimg`.