[vecz] Packetize sub-group shuffle_xor builtins
This extends fixed-width vectorization capabilities to
`__mux_sub_group_shuffle_xor` builtins. This isn't something that is
very efficiently vectorized, because of all of the runtime indexing,
which no built-in LLVM instructions/intrinsics can really make use of.

It might be preferable for some targets to go through memory. We might
want to make that a codegen option in a future update.