[doc] Update list of tested/supported LLVM versions
I've removed the distinction between the two targets, as we nominally
support both LLVM versions on each. A bug in either target in either
version of LLVM is valid to report, and as such it is more useful for
users to be aware of what we *claim* to support than what we actively
test. The distinction was more valid in an earlier time when riscv had
its own special patched version of LLVM.

Note that this has always been an overly simplified document; it's
impossible to convey exactly *what* we test in which configurations on
which targets and when, without just providing access to our internal
testing infrastructure.