[host] Fix the maximum number of sub-groups a kernel reports
This fixes a bug in the mux kernel query for the maximum number of
sub-groups.

We were mistakenly returning a number of sub-groups higher than the
maximum number of sub-groups reported by the device. The device
correctly reports that the theoretical maximum number of sub-groups
across all kernels was the maximum local work-group size.
However, the kernel query was multiplying the theoretical maximum number
of work-items in each of the separate X, Y, and Z dimensions together,
which may exceed the maximum work-group size.

We're relying on the fact that the OpenCL and SYCL specifications allow
this kind of flexible behaviour with regards to sub-group sizes (i.e.,
allowing the same kernel to exhibit different sub-group sizes depending
on the ND range). Re-addressing this would take a much wider redesign of
how we compile sub-groups, and we're not currently forced to by any
specification. Therefore this is a good enough fix for the short- to
medium-term.