[USM] Accept arbitrary pointers.
The SYCL CTS reveals that the logic we had in
clSetKernelArgMemPointerINTEL to detect whether the pointer arguments we
passed in were valid, while supported by the
cl_intel_unified_shared_memory specification, was causing valid SYCL
programs to be rejected, and the specification had an open question on
whether this is actually desired. This commit implements a change to
just accept arbitrary pointer values, which allows us to also remove a
bunch of USM special casing.