Use csv module for city runner
The SYCL CTS contains tests whose names contains commas. When we split
our CSV file on commas, this breaks down. Using Python's CSV parser
instead allows us to easily handle this.

The handling of comments and blank lines is left as is for now to
minimize the impact on CI.