[spirv-ll] Nominally support the DebugInfo/OpenCL.DebugInfo.100 sets
This adds a skeleton combined builder for these two sets. All of the
instructions are ignored, meaning they are not translated to LLVM IR.

Since both instruction sets only produce non-semantic instructions, this
is assumed to be safe enough for real-world usage, with a caveat noted
in the code that it's probably possible to create a legal SPIR-V binary
(from the verifier's point of view, at least) that reads the result of
one of these opcodes from another instruction set, at which point we may
internally error on not having produced an LLVM value for one a
DebugInfo opcode IDs.

This risk is assumed to be minimal enough for real-world SPIR-V modules
with debug info that nominally allowing the instruction set is more
beneficial than not.