[spirv-ll] Rename 'get' method for clarify
There are many overloaded 'get' functions in spirv-ll, and it can be
confusing whether we're getting an Op from a spv::Id, an Op from a
llvm::Value, or an Op from a llvm::Type. This is especially because
there are many variables we call variants of "ty" or "type":

* `spv::Id structType`
* `OpType *sourceOpType`
* `llvm::Type *resultType`

Tracking the types of all these variables can be difficult when many
things are declared with `auto`. One way of combatting this is by better
naming the functions which declare these types.