[spirv-ll] Fix and refactor parameter-type debug info
We were passing `align` to the third parameter of
`DIBuilder.createBasicType`, but that parameter is actually the
`Encoding`. This meant that we were generating debug information for
types which were, e.g., `float` for all 4-byte types or `unsigned_char`
for all 8-byte types. The align is only relevant for pointer types.

We also take the opportunity to refactor this method, taking in a ID
rather than an LLVM Type. This gives us more control over the type. For
example, LLVM pointer types won't be able to get their pointee types as
they're now opaque. Using the SPIR-V type ID we can better trace the
types.

This fixes another bug, as before, generating debug info type for
pointers was crashing as it tried to get the LLVM type for the pointee
type.