Fix exact multiplication of doubles.
The multiply_exact_helper<abacus_half> method has a comment explaining
that the non-specialized version is wrong for floating point types with
a mantissa with an odd number of bits. Like half, double also has an odd
number of bits in the mantissa, and we were generating incorrect
multiplication results as revealed by the SYCL CTS. So remove the
multiply_exact_helper<abacus_half> specialization and fix the primary
template definition instead.