Fix undefined behavior in is_odd.
We would previously evaluate mantissa << unbiasedExp and then try to
mask out the result based on validExp which tells us whether this shift
is safe to perform. For scalars, << goes by ordinary C++ rules which
say the behavior is undefined. For vectors, the specs are less clear and
this may be intended to be valid, but regardless, LLVM optimizes it the
same as scalar <<. The undefined behavior results in LLVM poison which
cannot be masked away. Mask the exponent as well so that we always have
a valid shift count.