[spirv-ll] Allow DILocs to span multiple ranges
The old code wouldn't work correctly when an OpLine was specified
multiple times with the same line/column, because we stored only one
range for each 'unique' OpLine.

This commit updates the translator to store a list of ranges for each
OpLine, meaning multiple distinct ranges can be attached to the same
debug location.