[spirv-ll] Validate SPIR-V glsl tests
See an earlier change to the spvasm for the rationale and methodology.

These tests are somewhat simpler than the spvasm tests as, in local
testing, glslangValidator never creates invalid SPIR-V binaries, and as
such all binaries are validated.

If this were ever to change, a similar system as used by the spvasm
tests could be employed.