[spirv-ll] Simplify FileCheck checks for variables
It's very rare that we actually care about the name of an LLVM value to
such a degree. As a translation tool using the LLVM libraries it'd be
very difficult for `spirv-ll` to generate invalid LLVM variable names.

The overly-specific regexes have been replaced with a much simpler one;
one that LLVM itself uses in its `update_*_checks.py` scripts.