Avoid memory leak in _cl_command_buffer_khr::updateCommandBuffer.
In _cl_command_buffer_khr::updateCommandBuffer we need to create a plain
old data descriptor, which requires storing the data externally. We have
no pre-existing memory which is guaranteed to persist for long enough
that we can use it, but we can add a field to UpdateInfo to hold it.