[spirv-ll] Remove deferred processing of OpPhi
Following the recent change to branches, we can remove the
deferred processing of Phis by:

1. Creating basic blocks on the fly if needed
2. Creating poison values if the value hasn't been visited yet, and
   finalizing that value later when it is visited

For the latter, this necessitates a change in how `addID` works. Now,
when a `addID` is called on an ID which already has a different recorded
value, the new value is assumed to be superior, and the old value is
replaced with the new. In practice this only came up once in testing
(exactly with phis) so I don't expect this to be an impactful change
except for this scenario.

This change to `addID` may open up changes for further simplification of
other forward-declared values.