[spirv-ll] Fix a debug info crash with BuiltIns
We were recording ranges of instructions to which to apply debug
information, but first we were replacing the initial IR for BuiltIn
variables with calls to the builtin functions. This meant that the
instruction ranges were sometime invalid, and the translator would
crash.

This commit fixes that issue by applying debug info to the original IR,
before we replace the BuiltIns.