[spirv-ll] Replace most internal uses of cargo types
This commit changes most internal uses of cargo data types:

* `cargo::optional<spirv_ll::Error>` -> `llvm:Error`
* `cargo::optional<T>` -> `std::optional<T>`
* `cargo::expected<T, std::string>` -> `llvm::Expected<T>`

The rationale is that `spirv-ll` is inherently an LLVM-based tool - we
don't build the tool at all in an LLVM-less build - and so making use of
more "standard" types should be more intuitive - returning
`llvm::Error::success()` instead of `cargo::nullopt`, for instance. We
can also make use of LLVM's more extensive support for dedicated error
handling, compared with our hand-rolled version.

I haven't changed the 'external' (or public) APIs just yet: those which
the `compiler` module makes use of. That can come later, if we so wish.