[mux][riscv] Fix getLocalSizeForSubGroupCount
We hadn't fully fleshed out this query on riscv. We were returning 0,0,0
(an invalid work-group size) for anything other than 1 sub-group. This
was being picked up as a failure by the OpenCL-CTS.

Instead, we must return a work-group size for which the work-group would
execute the same number of sub-groups as we've been asked for. The
returned work-group size is always 1D, hence we only care about the 'X'
dimension.

We do this by taking the maximum sub-group size we've compiled, and
returning the count multiplied by that size. This is because we know the
kernel variant picking algorithm favours the kernel variant with the
highest legal sub-group size. If the required work-group size is invalid
we must also report 0,0,0.

This fixes a test failure in the OpenCL CTS.

I've had to fix a bad UnitMux test which was incorrectly expecting the
returned local size to be X,1,1, where X couldn't be 1.