[spirv-ll] Attach names to values more often
We weren't doing a very thorough job at making use of OpNames when
translating SPIR-V to LLVM IR. This is obviously fine as the names are
purely decorative. However, they do have a use in making the translated
LLVM IR easier to intuit.

This is primarily because we were manually attaching names to the LLVM
IR instructions (or blocks, functions, etc.) when creating them. If we
forgot to do that, the names would be forgotten.

Instead, if we attach names centrally, we can achieve a far higher
coverage. We choose to do this in addID which is in practice called
during the translation of almost every SPIR-V opcode.

The name is only attached if one hasn't been given already. This means
that, while by default the name will be attached to the "last"
instruction in the translated sequence, this can be overridden if
desired. That said, since names are decorative, this isn't hugely
important and it's not often we translate opcodes using multiple
instructions. We can fix up any specific cases later if they're
particularly egregious.

This gives us the opportunity to more cleanly prioritze the names of
entry points over 'regular' functions. Previously we were doing this in
a more convoluted two-step process. We can instead do it on the fly.

This also fixes a bug where we could accidentally pick up GLSL functions
called 'printf' as the C-style printf, and incorrectly make it variadic.
A test has been added to cover it.