[spirv-ll] Add an interface for OpExtInstImport handlers
This goes part way towards making the handling of SPIR-V extended
instruction sets more flexible and less dependent on changing `spirv-ll`
itself.

By adding an abstract interface through which the handlers are
registered and are asked to handle a given `OpExtInst`, users of this
library may in the future be able to register their own external
handlers, if we so wish. This has been a long-standing wish for the
project.

In the short-term, it helpers to separate out almost all of the handling
of extended instruction sets into separate files. We were previously
defining all the handlers in the common `builder.h` header file.