[spirv-ll] Return debug strings as optional values
Knowing whether a debug string was actually provided is sometimes
useful, e.g., for error handling. By returning an empty string we are
unable to tell whether a string was provided and was empty, or whether
one wasn't provided.

Changing the return type of this function to
`std::optional<std::string>` can provide this info for us.