[spirv-ll] Associate debug subprograms with function IDs
Instead of mapping from each llvm::Function to its debug subprogram, we
want to instead associate the original SPIR-V ID representing that
function.

This is primarily to help the upcoming translation of the DebugInfo
extensions, as the `DebugFunction` instructions take a *forward
reference* to a function (meaning we only have its ID), and we want to
associate the translated subprogram with the function when the function
is finally processed later in the module.