[spirv-ll] Fully support OpenCL.DebugInfo.100 instructions
This commit extends spirv-ll to consume OpenCL.DebugInfo.100
instructions and translate them to meaningful LLVM debug info.

Although the legacy DebugInfo instruction set is similar, there are
differences, and we aren't seeing real-world binaries with this extended
instruction set and so lack tests to confidently enable it.

The translation process differs slightly from the translation of other
extended instruction sets. There is a small core of 'root' instructions
in the debug info hierarchy, and many more 'leaves' which don't
meaningfully add debug information on their own. Thus the translator
only processes the roots on demand as we visit them in program order.
Those roots may then reference other instructions (roots or leaves)
which we translate on the fly. We cache the translation of each
instruction as, in practice, they are usually referenced multiple times.

Generally, instructions are translated to `llvm::Metadata*`. The special
instruction `DebugInfoNone` translates to `nullptr`: this is not an
error; each instruction has to decide which operands may or may not be
`nullptr`. This usually ends up depending on what the LLVM APIs accept -
some accept `nullptr` as a valid value, others crash or assert.

One thing to note about the `OpenCL.DebugInfo.100` instruction set is
that there are numerous bugs all over the ecosystem:

* Even recent versions of `spirv-val` complain about what should be
  valid SPIR-V binaries
* `llvm-spirv` (and thus DPC++) contains known bugs that won't be fixed
  soon. We try to work around these as best as possible
  * Mixed instruction opcodes, explain below
  * Extra dummy operands which shouldn't be there
  * Forward references where none should be allowed
  * etc
* `llvm-spirv` (and thus DPC++) support undocumented extensions of their
  own, such as the ability to translate more LLVM expression opcodes
  than the spec permits.

The most egregious bug is that `llvm-spirv` has mixed up two instruction
opcodes and thus generates binaries which a standards-compliant SPIR-V
consumer will struggle with. To cope with this, we add a bitfield of
'workarounds' to the DebugInfoBuilder, which can toggle on and off
behaviour.

The only workaround added so far instructs the translator to, when faced
with one of these two instructions, try and infer which one is intended.
It does by inspecting the operands, using rules to establish whether the
instruction opcode is correct or is swapped with the other one. This
workaround is *enabled* by default, as it's still not fixed upstream. We
can toggle this behaviour off later, perhaps.