[spirv-ll] Provide helper methods for working with OpExtInst operands
Remembering and hard-coding the value '5' everywhere is brittle and can
be easy to miss when scanning code, leading to bugs. By introducing the
more intuitive notion of having operands of the extended instruction
modeled on top of the operands of the SPIR-V OpExtInst itself, we can
think at a higher level of abstraction.