Try to recover from fatal JIT errors.
On ARM, the following disabled UnitCL tests result in errors that take
down the whole UnitCL process:

  Execution/HalfOperatorTest.Precision_03_Half_Add/OpenCLC_2
  Execution/HalfOperatorTest.Precision_04_Half_Sub/OpenCLC_2
  Execution/HalfOperatorTest.Precision_05_Half_Mul/OpenCLC_2
  Execution/HalfOperatorTest.Precision_06_Half_Div/OpenCLC_2
  Execution/HalfOperatorTest.Precision_07_Half_Recip/OpenCLC_2

With this change, after those tests fail, UnitCL is able to continue
running to cover the remaining tests.

Note that we do fail an assertion at global test environment tear-down.
This happens because CrashRecoveryContext is implemented via
setjmp/longjmp, which is not required to call destructors for local
objects. Likely, this already resulted in memory leaks elsewhere, but
here, we leak a reference-counted string, and the string pool asserts
that no references remain in its destructor. This is not ideal, but I
see no good way around this, crashing there is more useful for us than
crashing right away, and in builds with assertions disabled, it will not
start doing the wrong thing.