[spirv-ll] Refactor how debug ranges/scopes are handled
This commit aims to simplify how we handle the management of different
types of debug scope.

Debug info is now attached to instructions when we close a range. We no
longer store all of the ranges and process them at the end of the
module. This keeps debug information better contained within the
builder, and means we have to track less volatile data. The code should
be simpler as a result, and hopefully easier to maintain.

We also introduce a new concept to help manage debug info. In addition
to the old 'line' range which is built in to SPIR-V, we introduce
another: a 'lexical scope'. This will be used for the translation of the
various DebugInfo extended instruction sets.

The lexical scopes and line ranges do interact, in the same way that the
DebugInfo instruction sets interact with the score spec: the DebugInfo
instruction sets still rely on line number information provided by line
ranges. A lexical scope is of no use without line information. We may
define lexical scopes in one of two ways, in priority order:

1. The DebugInfo extended instruction sets generate them using dedicated
   instructions
2. We generate them on the fly when attaching debug info when we process
   line ranges

Thus, when we close a line range or a lexical scope, we apply debug info
to all instructions within the range. For the scope information, we take
the lexical scope information, if set, and else we generate one on the
fly.