[refsi] Update Refsi memory specifications
In the refsi simulator, there are two memory regions:
* "Main" memory, starting at 0x10000
* "Local" memory, starting at 0x10000000

When emitting the sections, we move the output cursor to location
0x10000000 before writing the sections for local memory. This has
the effect of, in some unknown circumstances, resulting in the
generated ELF file including padding in the file itself for 255MiB.
Besides wasting memory, this also made the refsi simulator unable
to load the binaries due to its 128MiB limit.

This patch updates the linker script to be more explicit about the
memory layout, which also has the added advantage of having the
linker verify that the binary isn't too large.

It also reverts 73c6af20, which was a workaround for this issue.