Improve experience running tutorial script twice without '-o'
The previous behaviour would be that the `patch` commands would enter
`-R` mode and the symlink command would fail noisily.

It now provides a warning explaining what's happened: that the patch
files weren't applied (assumed to be already applied) and that the
symlink already exists, so doesn't need to be set.