Allow passing relative paths to create_target.py
Passing relative paths to `--external-dir` would do strange things.

I've observed it not creating all of the directories during the first
step, and one of the post-gen hooks would raise an exception (possibly
as a result of that).

The simplest fix is to ensure the external directory is made absolute
before working with it.