Fixed threading issues with usm and kernel execution
Usm adds some extra event handling so that any freeing of usm
memory is safe. There is some attempt already to make this thread
safe but this leaves windows where we have added events to the usm
allocations but we have not actually added the kernel execution to
the queue. This can mean the functions such as clMemBlockingFreeINTEL
can end up waiting on an event in a queue too early.

This resolves this by placing the context mutex across both the usm
event creation and the push to the queue, which means that the
usm deletion correctly waits.