Basic shared USM support
This implements shared allocations for USM using the same
machinery as host allocations. A shared USM allocation can be
associated with at most one device, and is accessable on host and
that device.

This should be compliant with the extension spec, but does not yet
support cross device access or migrating the allocation to the
device. Both `CL_MEM_ALLOC_INITIAL_PLACEMENT_DEVICE_INTEL` and
`clEnqueueMigrateMemINTEL` have no effect.