Minimal support for out-of-order queues
Out-of-order queues impose additional restrictions on user code and
grant additional freedom to the implementation, including the freedom to
continue to execute in order. Therefore, we can trivially support
out-of-order queues by removing the code that is specifically there to
reject them, and continuing to run them in order.

DPC++ aims to support OpenCL implementations that lack out of order
execution support, but it does so by creating temporary queues that
appear to not get released (intel/llvm#11156). By claiming out of order
execution support, we avoid leaks, and the specification for
CL_QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE only permits out-of-order
execution, it does not and in general cannot mandate it, so this should
be a valid minimal implementation.

In order to ease testing, this is optional for now, but enabled by
default. -DCA_ENABLE_OUT_OF_ORDER_EXEC_MODE=OFF can be used to restore
the prior behavior.

* CMakeLists.txt: add CA_ENABLE_OUT_OF_ORDER_EXEC_MODE option.
* config.h.in: report the CA_ENABLE_OUT_OF_ORDER_EXEC_MODE option.
* _cl_command_queue::create: tolerate
  CL_QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE and pass it along to
  _cl_command_queue, provided CA_ENABLE_OUT_OF_ORDER_EXEC_MODE is
  enabled.
* cl::EnqueueWaitForEvents: tolerate
  CL_QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE and ignore it, provided
  CA_ENABLE_OUT_OF_ORDER_EXEC_MODE is enabled.

Tests for unsupported properties are updated to test CL_QUEUE_ON_DEVICE
rather than CL_QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE.