[compiler] Improve diagnostics for failing sub-groups
Providing the name of the offending kernel is helpful when there is no
diagnostic information to point to a specific line.

It doesn't guarantee that the function means anything to the application
developer, as the failing function will (almost certainly) be mangled
and may have been wrapped by some of our other compiler passes. Even
then, this is still arguably useful information to present to people.