[host][riscv] Advertise/support a sub-group size of 32
Some SYCL benchmarks (seemingly ones ported to/from CUDA) require a
sub-group size of 32.

There shouldn't be any harm in us advertising this, even if the
vectorization width is larger than we typically consider:

* If it's required by a program and we don't support it, we'll fail to
  compile
* If we advertise it but fail to vectorize to that width, we'll fail to
  compile
* If we advertise it and successfully vectorize, then this is an
  improvement on what we had before

I've tried this with the `cudaSift` benchmark as part of Velocity Bench,
as we do successfully vectorize to this amount. I've noticed some poor
code-generation but we can work on improving that later.