[host] Fix LIT tests when CA_HOST_TARGET_CPU is set
The user-defined value was mistakenly passed to
`lit.util.pythonize_bool` which requires it to be a bool or a string
representing on/off: 'on','off','true','false', etc.

Since this is not a boolean option but a string representing a CPU,
there is a simpler safer way of performing this assignment.