Improve precision of horner_polynomial
We were previously implicitly relying on LLVM's behaviour of promoting
f16 operations to f32 and getting extra precision that amounts to
getting the same results as FMA. We never did that consistently (it
varied between platforms) and we now consistently do not do that,
resulting in incorrect results for at least half-precision pow. Use FMA
to fix that.

tanpi and tgamma have hardcoded exceptions where the approximation was
known to not be sufficiently precise. These exceptions are updated.