[NFC] Build Abacus in C++17 mode.
The previous -std=c++11 for Abacus dated from back when the default mode
was C++03, it was to enable new features in Abacus. The project-wide mode
now is C++17, and keeping -std=c++11 for Abacus now has the opposite
effect where features we can use throughout the project are not
available for Abacus. This change makes it so that the whole project
uses C++17.

This is a fairly large change because C++17 makes noexcept (including
__attribute__((nothrow))) part of the type system, requiring us to be
consistent between declaration and definition, we can no longer rely on
it being picked up from a previous declaration.