Avoid undefined behavior in FMA.
If the result is zero, the result is not denormal, so we do not need to
follow the denormal special case logic. The check
rounding_mask & ansAsUint would always be zero, so it would never do
anything, but the calculation of rounding_mask might have had undefined
behavior due to an out-of-range exponent.