Remove CA_CL_FORCE_PROFILE
This had been available as a backdoor for testing embedded but this
is untested and not documented and embedded is much lower priority
now. If needed we should make a target which is effectively embedded,
which is a much better way of testing it.

This also removes an unncessary compiler dependence on the cargo module
which would no longer compile if the CA_CL_FORCE_PROFILE_STRING was set
as a compile option.