USM testing rework for better USM support
This change brings a lot of duplicated code into
`cl_intel_unified_shared_memory_Test` (the base class of many
tests). Specifically, it provides properties to check host and
shared memory USM support, and helper functions to generate USM
pointers.

One of the design changes was adding a way to iterate through all
available USM pointer types without having to copy-paste code. This
is used to add shared allocations to many places alongside device
and host allocations.

This results in the tests doing more work, but this only causes a
10% reduction in the time taken for testing, which is about 300ms
on my system.