[spirv-ll] Add noundef param attrs to kernel entry points
In SPIR-V, entry points can only be called from outside the module, and
not from other functions inside the module.

Given this, we can safely add 'noundef' parameter attributes to entry
points with the "kernel" execution mode. Passing undefined or poison
values to a kernel is undefined behaviour. We could perhaps extend this
to other execution modes but we don't test those quite as well and they
aren't as important to us right now.

We'd ideally like to do this with *all* function parameters, which would
better match the semantics of C, C++, OpenCL C, but as noted in the
FIXME in the code, SPIR-V doesn't specify that passing an undefined
value to a function call results in undefined behaviour, so we need to
tread carefully. We might need to use `freeze` in more places to stop
the propagation of undef/poison values around the module.