Improve handling of error cases
Fix two issues I encountered while investigating something else.

Firstly, our tests don't initialize opencl events if they don't
expect a failure. While in many cases, calls consuming these events
shouldn't fail (and thus initialize the event correctly), it's
possible that, due to a bug, they fail resulting in the event
remaining uninitialized. If that event is then fed into functions
later in the test, it's better for behaviour to be defined than
undefined. We default initialise all opencl event pointers.

Secondly, if `MemFillImpl` gets into an unspecified state, it can
return `CL_INVALID_COMMAND_QUEUE`. However, the manner in which it
does this results in the event object not being properly tracked,
resulting in an event that never terminates yet is depended on by
an allocation. This results in a hang when deconstructing the
USM pointer. I don't know if this issue can occur when running
sound OpenCL code (it's possible if we don't have enough memory
to allocate a slot in the USM allocation's event list). We ensure
that the event is always marked as completed, even if the function
exits early due to an error.